+++ NEWS.md
@@ -17,6 +17,7 @@ They have been deprecated with warnings since v0.5.2
 #### Bug fixes
 * `mutate.()`: Can delete a column using `NULL` when `.by` is provided (#151)
 * `mutate_across.()`: Can reference other columns in .fns call (#145)
+* `slice.()`: Works when `.by` contains all columns (#158)
 * `unnest.()`: Works when the only column is the list column (#144)
 
 # tidytable 0.5.6